[Bugfix] Convert neuron ids to longs

+++ hbp_nrp_cle/hbp_nrp_cle/brainsim/pynn/__init__.py
@@ -56,4 +56,4 @@ class PyNNPopulationInfo(PopulationInfo):
         """
         Gets the global unique identifiers for this population
         """
-        return [i for i in self.__population.all()]
+        return [long(i) for i in self.__population.all()]
